Refactor single user mode
