+++ cmd/things/main.go
@@ -33,9 +33,9 @@ import (
 	authgrpcapi "github.com/mainflux/mainflux/things/api/auth/grpc"
 	authhttpapi "github.com/mainflux/mainflux/things/api/auth/http"
 	thhttpapi "github.com/mainflux/mainflux/things/api/things/http"
+	localusers "github.com/mainflux/mainflux/things/auth"
 	"github.com/mainflux/mainflux/things/postgres"
 	rediscache "github.com/mainflux/mainflux/things/redis"
-	localusers "github.com/mainflux/mainflux/things/users"
 	stdprometheus "github.com/prometheus/client_golang/prometheus"
 	jconfig "github.com/uber/jaeger-client-go/config"
 	"google.golang.org/grpc"
