+++ things/auth/doc.go
@@ -1,0 +1,8 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+
+// Package auth contains implementation for auth service in
+// single user scenario. Single user mode enables running
+// Things as a standalone service with a single admin user
+// that manages all the Things and Channels.
+package auth
